replace old findSimilar with findSimilarStructuredOutput.  Cleaned up cruft
